// src/types.ts
export type Animal = {
  id: string
  name: string
  image: string
  description: string
  origin: string
  colors: string[]
  category: string
  place_of_found?: string
  diet?: string
  habitat?: string
  family?: string
  species?: string
  temperament?: string
  size?: string
  lifespan?: string
}